requirements.txt
langchain-google-genai
google-generativeai
langchain_community
langchain
sentence-transformers
chromadb
python-dotenv
